+++ packages/react/src/components/Modal/ModalContent/index.ts
@@ -1,0 +1,2 @@
+import { ModalContent } from './ModaContent';
+export { ModalContent };
