+++ packages/react/src/components/Modal/ModalFooter/index.ts
@@ -1,0 +1,2 @@
+import { ModalFooter } from './ModalFooter';
+export { ModalFooter };
