+++ packages/react/src/components/Modal/ModalHeader/index.ts
@@ -1,0 +1,2 @@
+import { ModalHeader } from './ModalHeader';
+export { ModalHeader };
